feat: Implement custom color picker and refactor color management
Introduces a full-featured custom color picker dialog and refactors the application to use `androidx.compose.ui.graphics.Color` objects instead of color resource IDs for all decorations and UI elements. This allows for a much wider range of color choices and simplifies color state management.

This commit includes the following changes:
- Adds a new `ColorPickerDialog` composable using the `skydoves:colorpicker-compose` library.
- Integrates the color picker dialog, accessible via a new rainbow-colored button in the color selection rows.
- Replaces `Int` color resource IDs with `Color` objects across `Decoration`, `EditUiState`, `EditViewModel`, and repository layers.
- Updates `FansaUchiwaEntity` to store `Color` as a `Long` (ULong value).
- Implements `ColorSerializer` for kotlinx serialization and `ColorParceler` for `Parcelize` to handle `Color` objects.
- Updates `EditPager` and `EditScreen` to handle `Color` objects for decoration colors, uchiwa color, and background color.
- Modifies `DecorationColors` enum to hold `Color` values directly.
- Adds the `colorpicker-compose` dependency.

+++ app/src/main/java/com/example/fansauchiwa/data/ColorSerializer.kt
@@ -15,6 +15,6 @@ object ColorSerializer : KSerializer<Color> {
     }
 
     override fun deserialize(decoder: Decoder): Color {
-        return Color(decoder.decodeLong())
+        return Color(decoder.decodeLong().toULong())
     }
 }
